backend/requirements.txt
requests
python-dotenv
pandas
flask
flask-cors